requirements.txt
pandas>=2.2.0
python-dotenv>=1.0.1
google-api-python-client>=2.132.0
google-auth>=2.29.0
nltk>=3.9.1
tqdm>=4.66.4
requests>=2.32.3
# Optional: if you want to use Hugging Face sentiment later
# transformers>=4.44.2
# torch>=2.3.1
